% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_helpers.R
\name{check_exists}
\alias{check_exists}
\title{Check that element exists in a list}
\usage{
check_exists(list, name)
}
\arguments{
\item{list}{list to check}

\item{name}{element}
}
\description{
Check that element exists in a list
}
