% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_helpers.R
\name{check_whole}
\alias{check_whole}
\title{Check that an element exists and is a whole number}
\usage{
check_whole(list, name, required = TRUE)
}
\arguments{
\item{list}{list to check}

\item{name}{element}

\item{required}{whether the element is required or not}
}
\description{
Check that an element exists and is a whole number
}
