% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_validator.R
\name{forecast_validator}
\alias{forecast_validator}
\title{EFI forecast standard EML metadata validator}
\usage{
forecast_validator(eml)
}
\arguments{
\item{eml}{}
}
\value{
boolean
}
\description{
EFI forecast standard EML metadata validator
}
\examples{

forecast_validator(system.file("vignettes", "forecast-eml.xml",package="EFIstandards"))

}
