% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_validator.R
\name{forecast_validator}
\alias{forecast_validator}
\title{EFI forecast standard EML metadata validator}
\usage{
forecast_validator(eml)
}
\arguments{
\item{eml}{EML file path}
}
\value{
boolean
}
\description{
EFI forecast standard EML metadata validator
}
\examples{
\donttest{
forecast_validator(system.file("extdata", "forecast-eml.xml", package="EFIstandards"))
forecast_validator(system.file("extdata", "beetles-eml.xml", package="EFIstandards"))
}
}
